move eventstore test suite to base
