+++ Makefile
@@ -4,7 +4,7 @@ all:
 
 test:
 	#base
-	cd base && go test -count 1 validator_test.go
+	cd base && go test -count 1 ./...
 	# event stores
 	cd eventstore/bbolt && go test -count 1 ./...
 	cd eventstore/sql && go test -count 1 ./...
