+++ eventstore/bbolt/bbolt_test.go
@@ -5,8 +5,8 @@ import (
 	"testing"
 
 	"github.com/hallgren/eventsourcing/base"
+	"github.com/hallgren/eventsourcing/base/suite"
 	"github.com/hallgren/eventsourcing/eventstore/bbolt"
-	"github.com/hallgren/eventsourcing/eventstore/suite"
 )
 
 func TestSuite(t *testing.T) {
