+++ eventstore/memory/memory_test.go
@@ -4,8 +4,8 @@ import (
 	"testing"
 
 	"github.com/hallgren/eventsourcing/base"
+	"github.com/hallgren/eventsourcing/base/suite"
 	"github.com/hallgren/eventsourcing/eventstore/memory"
-	"github.com/hallgren/eventsourcing/eventstore/suite"
 )
 
 func TestSuite(t *testing.T) {
